refactor: Split monolithic modules and improve architecture (fixes #243)
Architecture improvements based on review findings:

1. Extract shared data-loader utility (src/services/data-loader.js)
   - Consolidates duplicate load-decrypt-store-emit pattern
   - Used by categories.js, contexts.js, priorities.js

2. Split todos.js (1133 lines) into focused modules
   - todos.js: Core CRUD + barrel re-exports for backward compat
   - todos-bulk.js: Bulk operations (delete, update status/project/priority)
   - todos-filters.js: getFilteredTodos, getGtdCount, getProjectTodoCount
   - todos-selection.js: Toggle, range, select all, clear selection
   - todos-recurrence.js: Create, generate, check, stop, convert recurring

3. Fix document-level event listener leaks in app.js
   - Add AbortController for global keyboard and click listeners
   - Clean up listeners on sign-out, re-register for next session

4. Add search input debouncing (150ms)
   - Prevents full re-render on every keystroke

5. Extract RecurrencePanel from TodoModal
   - TodoModal.js reduced from 805 to 423 lines
   - New RecurrencePanel.js handles all recurrence UI logic

All existing imports continue to work via barrel re-exports.

Fixes #243
