+++ src/services/priorities.js
@@ -1,25 +1,18 @@
-import { supabase } from '../core/supabase.js'
-import { store } from '../core/store.js'
-import { events, Events } from '../core/events.js'
+import { Events } from '../core/events.js'
+import { loadCollection, getById } from './data-loader.js'
 
 /**
  * Load all priorities for the current user
  * @returns {Promise<Array>} Array of priorities
  */
 export async function loadPriorities() {
-    const { data, error } = await supabase
-        .from('priorities')
-        .select('*')
-        .order('level', { ascending: true })
-
-    if (error) {
-        console.error('Error loading priorities:', error)
-        throw error
-    }
-
-    store.set('priorities', data)
-    events.emit(Events.PRIORITIES_LOADED, data)
-    return data
+    return loadCollection({
+        table: 'priorities',
+        storeKey: 'priorities',
+        event: Events.PRIORITIES_LOADED,
+        orderBy: 'level',
+        decryptFields: []
+    })
 }
 
 /**
@@ -28,6 +21,5 @@ export async function loadPriorities() {
  * @returns {Object|null} Priority object or null
  */
 export function getPriorityById(id) {
-    const priorities = store.get('priorities')
-    return priorities.find(p => p.id === id) || null
+    return getById('priorities', id)
 }
